refactor(BarChart): move the content to a specific file and some context
