+++ index.html
@@ -6,7 +6,13 @@
   </head>
 
   <body>
-    <div id="app"></div>
+    <h1>#1 Visualize Data with a Bar Chart</h1>
+    <a
+      href="https://www.freecodecamp.org/learn/data-visualization/data-visualization-projects/visualize-data-with-a-bar-chart"
+      target="_blank"
+      >User stories</a
+    >
+    <div id="barChart"></div>
     <script src="src/index.js"></script>
     <script src="https://cdn.freecodecamp.org/testable-projects-fcc/v1/bundle.js"></script>
   </body>
